Rename utils module to amqp utils.
