+++ test/unit/services/api/api_test.py
@@ -23,7 +23,7 @@ def test_client():
     ctx.pop()
 
 
-@patch('mash.services.api.routes.utils.Connection')
+@patch('mash.services.api.routes.amqp_utils.Connection')
 def test_api_add_account_ec2(mock_connection, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -62,7 +62,7 @@ def test_api_add_account_ec2(mock_connection, test_client):
     assert response.data == b'{"name":"test"}\n'
 
 
-@patch('mash.services.api.routes.utils.Connection')
+@patch('mash.services.api.routes.amqp_utils.Connection')
 def test_api_add_account_gce(mock_connection, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -109,7 +109,7 @@ def test_api_add_account_gce(mock_connection, test_client):
     assert response.data == b'{"name":"test"}\n'
 
 
-@patch('mash.services.api.routes.utils.Connection')
+@patch('mash.services.api.routes.amqp_utils.Connection')
 def test_api_add_account_azure(mock_connection, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -161,7 +161,7 @@ def test_api_add_account_azure(mock_connection, test_client):
     assert response.data == b'{"name":"test"}\n'
 
 
-@patch('mash.services.api.routes.utils.Connection')
+@patch('mash.services.api.routes.amqp_utils.Connection')
 def test_api_delete_account_ec2(mock_connection, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -193,7 +193,7 @@ def test_api_delete_account_ec2(mock_connection, test_client):
     assert response.data == b'{"name":"test"}\n'
 
 
-@patch('mash.services.api.routes.utils.Connection')
+@patch('mash.services.api.routes.amqp_utils.Connection')
 def test_api_delete_account_gce(mock_connection, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -225,7 +225,7 @@ def test_api_delete_account_gce(mock_connection, test_client):
     assert response.data == b'{"name":"test"}\n'
 
 
-@patch('mash.services.api.routes.utils.Connection')
+@patch('mash.services.api.routes.amqp_utils.Connection')
 def test_api_delete_account_azure(mock_connection, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -258,7 +258,7 @@ def test_api_delete_account_azure(mock_connection, test_client):
 
 
 @patch('mash.services.api.routes.jobs.uuid')
-@patch('mash.services.api.routes.utils.Connection')
+@patch('mash.services.api.routes.amqp_utils.Connection')
 def test_api_add_job_ec2(mock_connection, mock_uuid, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -297,7 +297,7 @@ def test_api_add_job_ec2(mock_connection, mock_uuid, test_client):
 
 
 @patch('mash.services.api.routes.jobs.uuid')
-@patch('mash.services.api.routes.utils.Connection')
+@patch('mash.services.api.routes.amqp_utils.Connection')
 def test_api_add_job_gce(mock_connection, mock_uuid, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -336,7 +336,7 @@ def test_api_add_job_gce(mock_connection, mock_uuid, test_client):
 
 
 @patch('mash.services.api.routes.jobs.uuid')
-@patch('mash.services.api.routes.utils.Connection')
+@patch('mash.services.api.routes.amqp_utils.Connection')
 def test_api_add_job_azure(mock_connection, mock_uuid, test_client):
     channel = MagicMock()
     connection = MagicMock()
@@ -374,7 +374,7 @@ def test_api_add_job_azure(mock_connection, mock_uuid, test_client):
         b'{"job_id": "12345678-1234-1234-1234-123456789012"}'
 
 
-@patch('mash.services.api.routes.utils.Connection')
+@patch('mash.services.api.routes.amqp_utils.Connection')
 def test_api_delete_job(mock_connection, test_client):
     channel = MagicMock()
     connection = MagicMock()
